# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TodoCamisetas API",
        "description": "API REST para gestionar camisetas, clientes, tallas y calcular precio final segun el cliente consultado.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Servidor de desarrollo local"
        }
    ],
    "paths": {
        "/camisetas": {
            "get": {
                "tags": [
                    "Camisetas"
                ],
                "summary": "Listar camisetas",
                "operationId": "getCamisetas",
                "parameters": [
                    {
                        "name": "club",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "pais",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tipo",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "color",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado exitoso"
                    }
                }
            },
            "post": {
                "tags": [
                    "Camisetas"
                ],
                "summary": "Crear camiseta",
                "operationId": "createCamiseta",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "titulo",
                                    "club",
                                    "pais",
                                    "tipo",
                                    "color",
                                    "precio",
                                    "codigo_producto"
                                ],
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "example": "Camiseta Local 2025"
                                    },
                                    "club": {
                                        "type": "string",
                                        "example": "Seleccion Chilena"
                                    },
                                    "pais": {
                                        "type": "string",
                                        "example": "Chile"
                                    },
                                    "tipo": {
                                        "type": "string",
                                        "example": "Local"
                                    },
                                    "color": {
                                        "type": "string",
                                        "example": "Rojo"
                                    },
                                    "precio": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 45000
                                    },
                                    "precio_oferta": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 39990,
                                        "nullable": true
                                    },
                                    "detalles": {
                                        "type": "string",
                                        "example": "Tela dry-fit, version hincha.",
                                        "nullable": true
                                    },
                                    "codigo_producto": {
                                        "type": "string",
                                        "example": "CHI-LOC-2025-01"
                                    },
                                    "cliente_id": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Camiseta creada"
                    },
                    "422": {
                        "description": "Errores de validacion"
                    }
                }
            }
        },
        "/camisetas/{id}": {
            "get": {
                "tags": [
                    "Camisetas"
                ],
                "summary": "Obtener camiseta por ID",
                "operationId": "getCamiseta",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "cliente_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Camiseta encontrada"
                    },
                    "404": {
                        "description": "Camiseta o cliente no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Camisetas"
                ],
                "summary": "Actualizar camiseta",
                "operationId": "updateCamiseta",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Camiseta actualizada"
                    },
                    "404": {
                        "description": "Camiseta no encontrada"
                    },
                    "422": {
                        "description": "Errores de validacion"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Camisetas"
                ],
                "summary": "Eliminar camiseta",
                "operationId": "deleteCamiseta",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Camiseta eliminada"
                    },
                    "404": {
                        "description": "Camiseta no encontrada"
                    }
                }
            }
        },
...
    },
...
}